requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
requests==2.31.0
python-dotenv==1.0.0
markdown2==2.4.10
pydantic==2.5.2
weasyprint==60.1
python-frontmatter==1.0.0
tenacity==8.2.3
tqdm==4.66.1
